requirements.txt
django>=1.11,<1.12
django-debug-toolbar>=1.7
gunicorn>=19.7.0
gevent>=1.2.2
psycopg2-binary>=2.7.1
whitenoise>=3.3
django-model-utils>=2.6.1
django-crispy-forms>=1.6.1
django-formtools>=2.0
django-settings-export>=1.2.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
minio>=3.0.0
nose>=1.3.7
django-nose>=1.4.5
coverage>=4.4.2
django-webpack-loader==0.6.0
django-filter==1.1.0
drf-yasg==1.5.0
